chore: move PushAnalyticsError to dir and add parseError helper function
