+++ src/PrimaryProcess/DashboardItemQueue.ts
@@ -1,4 +1,4 @@
-import { PushAnalyticsError } from '../PushAnalyticsError'
+import { PushAnalyticsError } from '../Error'
 import type { AddDashboardOptions, QueueItem } from '../types'
 
 class DashboardItemsQueueError extends PushAnalyticsError {
